backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
redis
celery
pydantic[email]
passlib[bcrypt]
bcrypt==3.2.2
passlib==1.7.4
python-jose[cryptography]
python-multipart
alembic
python-docx
markdown
pyyaml
